renamed single ticket generator files
